start on magic-wormhole in Rust
I'm not sure how the repos ought to be split yet. I'd like to preserve this
one for a CLI tool, so "cargo install magic-wormhole; wormhole send FILE"
works. But I don't want applications which use the wormhole library to be
forced to include CLI libraries as dependencies. So the bulk of the code
might get split out to a magic-wormhole-lib package, leaving this one with
just the CLI layer.