boss: fix close()
It should tell the Terminator to begin shutdown, and then wait for the
Terminator to report completion. In particular the Terminator is the one
which shuts down the Rendezvous connection, not the Boss.

By waiting to free the mailbox and nameplate, we also ensure that all other
messages are delivered, such as regular data messages. This allows an
application to do `w.send(msg) ; w.close()` and know that the message will be
queued before the connection is dropped, without waiting for an
application-level ack from the peer.