add MySide newtype
This wraps a String, and is used in all places where we send our "side" value
to the server. This is distinct from the kind of string that we receive from
the other client, although there is one place (in Mailbox) where we compare
them to see if a given message is a reflection of our own, or a new one from
a peer.