add Code newtype
This wraps a String.

This doesn't compile because it highlights a bug: the code and appid
arguments were swapped going into start_pake(). The use of distinct types
catches this, which is great because swapping those arguments breaks the
security of SPAKE2 (it enables an online man-in-the-middle attack that only
requires testing one hash per possible password value).

I'll fix that bug in the next commit.