server_messages: fix Mailbox serialization
By using Display and .to_string() on the Mailbox newtype, our
wire-serialization picked up a bogus "Mailbox()" wrapper, which wasn't fatal
but prevents the server from correctly cleaning up (the server allocated one
mailbox for us, then we did an OPEN of a differently-named mailbox.. we
released the second one but not the first).

This fixes that and changes the tests to parse the output as JSON rather than
using the deserialization half of our outbound serializers, which I think
masks the problem.