start on Timing
This is a performance tool: various internal operations generate "Events"
with timestamps, and the list of events can be written out as a JSON file.
We've got visualization tools in the Python repository to render these in a
web browser, to figure out where we're spending a lot of time waiting for
roundtrips, to point at possible improvements. The Events can be merged from
both sides of the connection, as well as the mailbox server, and the viz tool
shows all the messages as they traverse the two connections.