return Timing events up to core
This produces log events for PAKE (in key.rs), the input_code span in
input.rs, and the receipt of messages in rendezvous.rs. Outbound rendezvous
messages are partially logged.

The Python version also emits logs for some points in Tor, Transit, and
Dilation, which we haven't ported yet. The same is true for the large number
of events in the Python-side CLI, which we haven't written yet.