simplify a state machine by using Option<State>
By using .take() to remove the old state from our struct, we can match
against it (and assign to self) without conflicts. The down side is that we
must always restore the state when we're done, so I have the main match()
clause return the new state (which might be a reconstructed version of the
old one).

I'm curious about whether this causes a slowdown in the common case: once we
move into S1HaveVerifiedKey, we stay there forever, but this approach means
we have to build a new state object every time (each one being identical to
the one before it).

But I have to admit the code is simpler this way.