Simplify startup and a lot of state machines
The `Start` event got removed. All state machines that previously had to be started are created in their "already starting/connecting" phase.
This started a cascade of simplifications where a lot of states in a lot of state machines could be eliminated.
The only price we pay for this is that some initialization has to be made upfront (in the constructor).
As we cannot reconnect anymore, there's only one WebSocket connection. We don't need the `WSHandle`s anymore, nor the ID management or retry timers.