Merge a *lot* of state machines
After all those simplifications, 'allocator' and 'code' suddenly had almost exactly the same states. 'code' was effectively calling 'allocator' as a sub-state-machine.
But because 'allocator' only had one non-trivial state, it could be inlined into 'code' without making it more complex.

Similar things happened to 'order', 'receive', 'code', 'terminator' and 'boss'.

The new code for terminating doesn't fully work yet. The event loop will panic while closing because the events are executed in the wrong order.
This does not really affect the functionality of the application, therefore it will be fixed later on.