Transfer refactoring, start of transfer-v2
This moves us towards the unified API that will be used in the future for
both v1 and v2 transfers. The latter already has bits of code, but it is mostly
a prototype and currently disabled.

Additionally, this commit gives great improvements to the cancellation
handling; the improved code for that has been moved out into its own submodule.